+++ templates/graph.php
@@ -27,20 +27,36 @@ include($_SERVER['DOCUMENT_ROOT']."/header.php");
 ?>
 <div class='settingListCard'>
 <?php
+if (array_key_exists("timespan",$_GET)){
+	$timespan=($_GET['timespan']);
+	echo "<h1>".ucfirst(file_get_contents("title.cfg"))." - ".ucfirst($timespan)."</h1>";
+}else{
 	echo "<h1>".ucfirst(file_get_contents("title.cfg"))."</h1>";
+}
+?>
+
+<?php
+if (array_key_exists("timespan",$_GET)){
+	echo "<a class='graphLink' href='$graph-$timespan.png'>";
+	echo "<img src='$graph-$timespan.png'>";
+	echo "</a>";
+}else{
 	echo "<a class='graphLink' href='$graph-day.png'>";
 	echo "<img src='$graph-day.png'>";
 	echo "</a>";
-	echo "<a class='graphLink' href='$graph-week.png'>";
-	echo "<img src='$graph-week.png'>";
-	echo "</a>";
-	echo "<a class='graphLink' href='$graph-month.png'>";
-	echo "<img src='$graph-month.png'>";
-	echo "</a>";
-	echo "<a class='graphLink' href='$graph-year.png'>";
-	echo "<img src='$graph-year.png'>";
-	echo "</a>";
+}
 ?>
+<div class='titleCard'>
+	<h2></h2>
+	<div class='listCard'>
+		<h2></h2>
+		<a class='button' href='?timespan=day'>Day</a>
+		<a class='button' href='?timespan=week'>Week</a>
+		<a class='button' href='?timespan=month'>Month</a>
+		<a class='button' href='?timespan=year'>Year</a>
+	</div>
+</div>
+
 </div>
 <?php
 	// add random comics above the footer
